backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
asyncpg>=0.29.0
python-dotenv>=1.0.1
pydantic>=2.7.0
httpx>=0.27.0
bleach>=6.1.0
